move functions to general
