+++ Yolo/utils.hpp
@@ -3,11 +3,6 @@
 #include <vector>
 #include <opencv2/core.hpp>
 
-enum strategy {
-    concatenatedContour = 0,    //concatenate all segments
-    largestContour = 1          //select largest segment
-};
-
 namespace Yolo {
 class Detector;
 struct Object;
@@ -66,12 +61,5 @@ private:
     Detector* detector;
     std::vector<std::string> class_names;
 
-private:
-
-    std::vector<cv::Point> mask2segment(const cv::Mat& mask, int strategy = largestContour);
-
-    cv::Mat applyMask(const cv::Mat& bgr, const cv::Mat& mask);
-
-    float getRotatedRectImg(const cv::Mat& src, cv::Mat& dst, const cv::RotatedRect& rr);
 };
 }
